Add Review API
Change recipe rating datatype.
Add APIs to Get, Create and Update recipe review.

+++ KitchenDelights/Data/Entity/Recipe.cs
@@ -17,7 +17,7 @@ public partial class Recipe
 
     public string? RecipeContent { get; set; }
 
-    public int RecipeRating { get; set; }
+    public decimal RecipeRating { get; set; }
 
     public bool RecipeStatus { get; set; }
